requirements.txt
django
scrapy
psycopg2-binary
sqlalchemy
python-dotenv
Pillow
transformers
ollama